+++ client/src/components/Main/FlipCard/SpaceInfoBack/style.ts
