+++ client/src/components/Main/FlipCard/SpaceSelectBack/style.ts
