+++ client/src/components/Main/FlipCard/style.ts
