+++ client/src/components/Main/PostMap/components/style.ts
