+++ client/src/components/Main/PostMap/style.ts
