+++ client/src/components/Main/SearchBox/style.ts
